Fixed Multi-System Piping
When I added counts, I broke the script when piping multiple system
names. I've added a band-aid to fix it (in the form of a System property
to the output object - can now us format-table -groupby System). Still
not perfect and I'll be re-working the object creation to simplify the
whole thing.

I also added a check that the Active Directory module is available.

Last bit - cleaned up some write-verbose code that was unnecessary.